+++ src/som/interpreter/nodes/OperationNode.java
@@ -3,4 +3,5 @@ package som.interpreter.nodes;
 
 public interface OperationNode {
   String getOperation();
+  int getNumArguments();
 }